\name{@ANNOBJPREFIX@PFAM}
\alias{@ANNOBJPREFIX@PFAM}
\title{Maps between Manufacturer Identifiers and PFAM Identifiers}
\description{
   @ANNOBJPREFIX@PFAM is an R object that provides mappings between
   manufacturer identifiers and PFAM identifiers.
}

\details{
The bimap interface for PFAM is defunct. Please use select() interface to
PFAM identifiers. See ?AnnotationDbi::select for details.
}

\keyword{datasets}
